easytech/requirements.txt
Django==5.1.7
djangorestframework==3.15.0
django-filter==24.1
django-admin-interface==0.29.4
django-colorfield==0.12.0
Pillow==11.1.0
python-slugify==8.0.4 